+++ TeamCode/src/main/java/teamcode/Commands/TurnAngleCommand.java
@@ -1,23 +1,21 @@
-package com.example.ftclibexamples.OldCommandSample;
+package teamcode.Commands;
 
-import com.arcrobotics.ftclib.command.old.Command;
-import com.arcrobotics.ftclib.controller.PController;
-
+import ftclib.command.old.Command;
+import ftclib.controller.PController;
 import org.firstinspires.ftc.robotcore.external.Telemetry;
+import teamcode.Subsystems.DriveSubsystem;
 
 public class TurnAngleCommand implements Command {
 
     DriveSubsystem driveSubsystem;
-    Telemetry tl;
     double angle;
 
     // Proportional Controller for correcting for gyro error
     PController headingController;
 
-    public TurnAngleCommand(DriveSubsystem driveSubsystem, double angle, Telemetry telemetry) {
+    public TurnAngleCommand(double angle, DriveSubsystem driveSubsystem) {
         this.driveSubsystem = driveSubsystem;
         this.angle = angle;
-        this.tl = telemetry;
         // At 180 degrees, we should spin almost as fast as we can to correct
         // 1 is full power. 180 * 0.05 = 0.9
         headingController = new PController(0.05, angle, driveSubsystem.getHeading());
@@ -29,13 +27,10 @@ public class TurnAngleCommand implements Command {
     @Override
     public void initialize() {
         // Reset gyro and encoders
-        driveSubsystem.reset();
+        driveSubsystem.resetEncoders();
 
-        // Set target to the target angle
-        tl.addData("Heading Setpoint", headingController.getSetPoint());
         // If within 5 degrees of setpoint, the target is considered reached
-        headingController.setTolerance(1);
-
+        headingController.setTolerance(5);
     }
 
     @Override
@@ -44,21 +39,17 @@ public class TurnAngleCommand implements Command {
         double rotate = headingController.calculate(driveSubsystem.getHeading());
 
         // apply output
-        driveSubsystem.driveTrain.driveRobotCentric(0, 0, rotate);
+        driveSubsystem.getRobotDrive().driveRobotCentric(0, 0, rotate);
     }
 
     @Override
     public void end() {
-        driveSubsystem.driveTrain.driveRobotCentric(0, 0, 0);
+        driveSubsystem.getRobotDrive().driveRobotCentric(0, 0, 0);
 
     }
 
     @Override
     public boolean isFinished() {
-        tl.addData("Position Error: ", headingController.getPositionError());
-        tl.addData("Heading Setpoint", headingController.getSetPoint());
-        tl.addData("At Setpoint", headingController.atSetPoint());
-        tl.addData("Current heading", driveSubsystem.getHeading());
         boolean angleReached = headingController.atSetPoint();
         return angleReached;
     }
